+++ src/output/json.cpp
@@ -401,6 +401,23 @@ void JsonOutput::printf(const std::string &str)
   emit_data(out_, "printf", std::nullopt, str);
 }
 
+void JsonOutput::errorf(const std::string &str, const SourceInfo &info)
+{
+  out_ << R"({"type": "errorf")";
+  out_ << R"(, "msg": )";
+  std::stringstream ss;
+  ss << str;
+  JsonEmitter<std::string>::emit(out_, ss.str());
+  // Json only prints the top level location
+  out_ << R"(, "filename": )";
+  JsonEmitter<std::string>::emit(out_, info.locations.begin()->filename);
+  out_ << R"(, "line": )";
+  JsonEmitter<uint64_t>::emit(out_, info.locations.begin()->line);
+  out_ << R"(, "col": )";
+  JsonEmitter<uint64_t>::emit(out_, info.locations.begin()->column);
+  out_ << R"(})" << std::endl;
+}
+
 void JsonOutput::time(const std::string &time)
 {
   emit_data(out_, "time", std::nullopt, time);