+++ src/output/json.h
@@ -13,6 +13,7 @@ public:
   void map(const std::string &name, const Value &value) override;
   void value(const Value &value) override;
   void printf(const std::string &str) override;
+  void errorf(const std::string &str, const SourceInfo &info) override;
   void time(const std::string &time) override;
   void cat(const std::string &cat) override;
   void join(const std::string &join) override;